Switch effect_name_chars to uint8_t to save space
It only takes on values 1 and 2.

+++ src/doc.h
@@ -148,7 +148,7 @@ struct DocumentCopy {
     /// Whether effect names are 1 or 2 characters wide.
     /// When set to 1, the first digit is hidden if it's 0,
     /// and typing character c will write effect 0c immediately.
-    uint32_t effect_name_chars = 1;
+    uint8_t effect_name_chars = 1;
 
     Samples samples;
     Instruments instruments;
